+++ ntex/src/http/h1/decoder.rs
@@ -630,7 +630,7 @@ impl ChunkedState {
 
         match size.checked_mul(16) {
             Some(n) => {
-                *size = n as u64;
+                *size = n;
                 *size += rem as u64;
 
                 Poll::Ready(Ok(ChunkedState::Size))
